\name{PerfMeas-package}
\alias{PerfMeas-package}
\alias{PerfMeas}
\docType{package}
\title{
PerfMeas: Performance Measures for ranking and classification tasks
}
\description{
Metrics for ranking and classification tasks: Area Under the ROC Curve (AUC), F-scores, and precision at given recall level are implemented.
}
\details{
\tabular{ll}{
Package: \tab PerfMeas\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2014-09-07\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
This package implements a set of functions to estimate the AUC, F-score, precision, recall, specificity, accuracy according to the 0/1 loss, and precision at given recall level for ranking and classification problems.

Functions to compute the above measures for single classes or for sets of classes are provided. 
}
\author{
\emph{Giorgio Valentini} and \emph{Matteo Re}

DI, Dipartimento di Informatica

Universita' degli Studi di Milano

\email{\{valentini,re\}@di.unimi.it}

Maintainer: 
\emph{Giorgio Valentini} 
}


\keyword{ package }

