\name{PerfMeas-package}
\alias{PerfMeas-package}
\alias{PerfMeas}
\docType{package}
\title{
Performance Measures for Ranking and Classification Tasks
}
\description{
Metrics for ranking and classification tasks: Area Under the Receiving Operating Characteristic Curve (AUROC), Area Under the Precision and Recall Curve (AUPRC), F-scores, and precision at given recall level are implemented.
}
\details{
\tabular{ll}{
Package: \tab PerfMeas\cr
Type: \tab Package\cr
Version: \tab 1.2.5\cr
Date: \tab 2022-09-09\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
This package implements a set of functions to estimate the AUC, F-score, precision, recall, specificity, accuracy according to the 0/1 loss, and precision at given recall level for ranking and classification problems.

Functions to compute the above measures for single classes or for sets of classes are provided. 
}
\author{
\emph{Giorgio Valentini}

DI, Dipartimento di Informatica

Universita' degli Studi di Milano

\email{\{valentini\}@di.unimi.it}

Maintainer: 
\emph{Giorgio Valentini} 
}


\keyword{ package }

